# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Herstory AI - Backend REST API Specification",
    "description": "Herstory AI 백엔드 시스템 OpenAPI 3.0 명세서입니다.\n- Auth API: 로그인 & VIP 허브\n- Journey API: OCR 스캔, 기후 분석, Live Card 위젯, Apple Wallet .pkpass\n- Cart API: 스마트 장바구니 & ChoiceFit 피팅 신청\n- Store API: 면세점 체크인 & 매장 직원 태블릿 SSE 실시간 알림\n- Order API: 면세 결제 & Herstory Miles 적립\n- Care API: 현지 럭셔리 부티크 스팟, Google Maps, OpenAI 가죽 케어, 시티 스탬프, FCM\n- Flight API: 항공편 실시간 운항 정보 조회\n- Health API: 시스템 헬스 모니터링",
    "version": "v1.0.0",
    "contact": {
      "name": "Herstory Tech Team",
      "email": "tech@herstory.com"
    }
  },
  "servers": [
    {
      "url": "https://herstory-backend.onrender.com",
      "description": "Render Production Server (Live)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/journey/{journeyId}": {
      "get": {
        "tags": ["2. Journey API (여정 & 큐레이션)"],
        "summary": "여정 기본 상세 단건 조회",
        "description": "여정 ID를 통해 PNR, 출발/도착 공항, 출발 일시, 운항 상태 및 기후 분석 정보를 조회합니다.",
        "operationId": "getJourney",
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "description": "조회할 여정 ID",
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          },
          "400": {
            "description": "여정을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journey/scan": {
      "post": {
        "tags": ["2. Journey API (여정 & 큐레이션)"],
        "summary": "보딩패스 OCR 스캔 & 여정 등록",
        "description": "탑승권 OCR 스캔을 통해 PNR을 추출하고 비행 여정을 저장합니다.",
        "operationId": "scanBoardingPass",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journey/analysis/{journeyId}": {
      "get": {
        "tags": ["2. Journey API (여정 & 큐레이션)"],
        "summary": "목적지 기후 및 여행 분석 데이터 조회",
        "description": "목적지 날씨 및 기후 정보를 분석하여 맞춤형 럭셔리 상품을 추천합니다.",
        "operationId": "analyzeJourney",
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/journey/apple-wallet-pass/{journeyId}": {
      "get": {
        "tags": ["2. Journey API (여정 & 큐레이션)"],
        "summary": "SCR-201 Apple Wallet 디지털 패스(.pkpass) 메타데이터 생성",
        "description": "비행 탑승권 및 VIP 여정 보딩패스를 Apple Wallet PKPass 형태로 생성합니다.",
        "operationId": "getAppleWalletPass",
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppleWalletPassResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journey/apple-wallet-pass/download/{journeyId}": {
      "get": {
        "tags": ["2. Journey API (여정 & 큐레이션)"],
        "summary": "Apple Wallet (.pkpass) 바이너리 파일 다운로드",
        "description": "iOS 아이폰 디바이스 지갑(Wallet) 앱에 직접 추가할 수 있는 .pkpass 패스 파일을 다운로드합니다.",
        "operationId": "downloadAppleWalletPass",
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공 (바이너리 .pkpass 스트림)",
            "content": {
              "application/vnd.apple.pkpass": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journey/apple-wallet-pass/download-file/{journeyId}": {
      "get": {
        "tags": ["2. Journey API (여정 & 큐레이션)"],
        "summary": "아이폰 범용 패스 파일 다운로드 (Safari 다운로드 호환)",
        "description": "iOS Safari 보안 차단 없이 아이폰 파일 앱(Downloads)으로 직접 다운로드할 수 있는 패스 파일 API입니다.",
        "operationId": "downloadAppleWalletPassFile",
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공 (바이너리 스트림)",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cart/add": {
      "post": {
        "tags": ["3. Cart API (스마트 장바구니 & ChoiceFit)"],
        "summary": "스마트 장바구니 상품 추가",
        "description": "추천 상품을 스마트 장바구니에 담습니다.",
        "operationId": "addToCart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cart/choice-fit": {
      "put": {
        "tags": ["3. Cart API (스마트 장바구니 & ChoiceFit)"],
        "summary": "ChoiceFit (VIP 피팅 신청 여부) 상태 업데이트",
        "description": "장바구니 담기 후 피팅 신청 여부(true/false) 상태를 업데이트합니다.",
        "operationId": "updateChoiceFit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChoiceFitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cart/my": {
      "get": {
        "tags": ["3. Cart API (스마트 장바구니 & ChoiceFit)"],
        "summary": "내 스마트 장바구니 조회",
        "description": "현재 회원의 활성화된 스마트 장바구니 및 품목 목록을 조회합니다.",
        "operationId": "getMyCart",
        "parameters": [
          {
            "name": "memberId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/care/visetos-spots": {
      "get": {
        "tags": ["6. Care API (현지 부티크 스팟 & 케어)"],
        "summary": "현지 글로벌 럭셔리 매장 스팟 탐색 및 가죽 케어 메시지 조회",
        "description": "목적지의 현지 럭셔리 플래그십 스토어 및 Care Desk 정보를 조회하고 맞춤형 케어 가이드를 반환합니다.",
        "operationId": "getVisetosSpots",
        "parameters": [
          {
            "name": "memberId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FlightStatus": {
        "type": "string",
        "enum": ["SCHEDULED", "BOARDING", "COMPLETED", "CANCELLED"]
      },
      "CartStatus": {
        "type": "string",
        "enum": ["IN_CART", "CHECKED_OUT", "CANCELLED"]
      },
...
      "ProductCategory": {
        "type": "string",
        "enum": ["WATERPROOF", "BACKPACK", "TRAVEL_BAG", "ACCESSORY", "LEATHER_CARE"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer", "example": 400 },
          "error": { "type": "string", "example": "Bad Request" },
          "message": { "type": "string", "example": "해당 회원을 찾을 수 없습니다." },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
      "ScanRequest": {
        "type": "object",
        "required": ["memberId"],
        "properties": {
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "pnr": { "type": "string", "example": "HST999" },
          "rawOcrText": { "type": "string", "example": "BOARDING PASS PASSENGER: KIM/NOMAD FLIGHT: KE651 ICN BKK" },
          "origin": { "type": "string", "example": "ICN" },
          "destination": { "type": "string", "example": "BKK" }
        }
      },
      "ScanResponse": {
        "type": "object",
        "properties": {
          "journeyId": { "type": "integer", "format": "int64", "example": 1 },
          "pnr": { "type": "string", "example": "HST999" },
          "origin": { "type": "string", "example": "ICN" },
          "destination": { "type": "string", "example": "BKK" },
          "departureDateTime": { "type": "string", "format": "date-time" },
          "flightStatus": { "$ref": "#/components/schemas/FlightStatus" },
          "message": { "type": "string", "example": "보딩패스 OCR 스캔이 성공적으로 완료되었습니다." }
        }
      },
      "JourneyResponse": {
        "type": "object",
        "properties": {
          "journeyId": { "type": "integer", "format": "int64", "example": 1 },
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "memberName": { "type": "string", "example": "김노마드 (VIP)" },
          "pnr": { "type": "string", "example": "HST999" },
          "origin": { "type": "string", "example": "ICN" },
          "destination": { "type": "string", "example": "BKK" },
          "departureDateTime": { "type": "string", "format": "date-time" },
          "flightStatus": { "$ref": "#/components/schemas/FlightStatus" },
          "destinationWeather": { "type": "string", "example": "열대성 스콜 (기온 32°C, 습도 85%)" },
          "recommendationReason": { "type": "string", "example": "목적지 비/습도 기후에 적합한 럭셔리 방수 레더 컬렉션 맞춤 제안" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "example": 1 },
          "name": { "type": "string", "example": "럭셔리 클래식 방수 스타크 백팩" },
          "category": { "$ref": "#/components/schemas/ProductCategory" },
          "price": { "type": "number", "example": 1250000.00 },
          "stock": { "type": "integer", "example": 45 },
          "imageUrl": { "type": "string", "example": "https://images.unsplash.com/photo-1553062407-98eeb64c6a62?auto=format&fit=crop&w=800&q=80" },
          "description": { "type": "string", "example": "열대 스콜 방수 코팅 캔버스 백팩" },
          "isVipExclusive": { "type": "boolean", "example": false }
        }
      },
      "JourneyAnalysisResponse": {
        "type": "object",
        "properties": {
          "journeyId": { "type": "integer", "format": "int64", "example": 1 },
          "destination": { "type": "string", "example": "BKK (방콕 수완나품)" },
          "weatherInfo": { "type": "string", "example": "기온 32.5°C, 습도 88%" },
          "climateSummary": { "type": "string", "example": "고온 다습한 열대 몬순 기후" },
          "recommendationReason": { "type": "string", "example": "방수 에센셜 라인업 제안" },
          "recommendedProducts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Product" }
          }
        }
      },
...
      "AppleWalletPassResponse": {
        "type": "object",
        "properties": {
          "passTypeIdentifier": { "type": "string", "example": "pass.com.herstory.passport" },
          "serialNumber": { "type": "string", "example": "HERSTORY-PASS-HST999-1" },
          "teamIdentifier": { "type": "string", "example": "HST99PASSKIT" },
          "organizationName": { "type": "string", "example": "Herstory" },
          "description": { "type": "string", "example": "Herstory VIP Pass" },
          "logoText": { "type": "string", "example": "HERSTORY AI" },
          "boardingPassDetails": { "type": "object" },
          "pkpassDownloadUrl": { "type": "string", "example": "/api/v1/journey/apple-wallet-pass/download/HST999.pkpass" }
        }
      },
      "AddItemRequest": {
        "type": "object",
        "required": ["memberId", "productId"],
        "properties": {
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "productId": { "type": "integer", "format": "int64", "example": 1 },
          "quantity": { "type": "integer", "default": 1, "example": 1 }
        }
      },
      "ChoiceFitRequest": {
        "type": "object",
        "required": ["memberId", "choiceFit"],
        "properties": {
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "choiceFit": { "type": "boolean", "example": true }
        }
      },
      "ItemDetail": {
        "type": "object",
        "properties": {
          "cartItemId": { "type": "integer", "format": "int64", "example": 1 },
          "productId": { "type": "integer", "format": "int64", "example": 1 },
          "productName": { "type": "string", "example": "럭셔리 클래식 방수 스타크 백팩" },
          "category": { "type": "string", "example": "BACKPACK" },
          "price": { "type": "number", "example": 1250000.00 },
          "quantity": { "type": "integer", "example": 1 },
          "imageUrl": { "type": "string", "example": "https://images.unsplash.com/photo-1553062407-98eeb64c6a62?auto=format&fit=crop&w=800&q=80" }
        }
      },
      "CartResponse": {
        "type": "object",
        "properties": {
          "cartId": { "type": "integer", "format": "int64", "example": 1 },
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "choiceFit": { "type": "boolean", "example": true },
          "status": { "$ref": "#/components/schemas/CartStatus" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ItemDetail" }
          },
          "totalPrice": { "type": "number", "example": 1250000.00 }
        }
      },
...
      "VisetosSpot": {
        "type": "object",
        "properties": {
          "spotName": { "type": "string", "example": "Herstory Siam Paragon Flagship" },
          "address": { "type": "string", "example": "991 Rama I Rd, Bangkok" },
          "locationType": { "type": "string", "example": "Herstory Flagship & VIP Lounge" },
          "latitude": { "type": "number", "example": 13.7466 },
          "longitude": { "type": "number", "example": 100.5349 },
          "careServiceAvailable": { "type": "string", "example": "가죽 방수 코팅 케어 제공" }
        }
      },
      "CareResponse": {
        "type": "object",
        "properties": {
          "destination": { "type": "string", "example": "Bangkok" },
          "pushNotificationMessage": { "type": "string", "example": "방콕 도착 환영 가이드" },
          "visetosSpots": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/VisetosSpot" }
          }
        }
      },
...
    }
  }
}